backend/requirements.txt
flask
flask-cors
python-dotenv
openai
gunicorn
sqlite3
spacy
underthesea